python_service/requirements.txt
fastapi
uvicorn
pymongo
pymongo
google-generativeai
python-dotenv
pydantic
python-dotenv
pydantic
